requirements.txt
# The purpose of this file is to store *all* the dependencies
# setup.cfg
anndata
numpy
interrogate
scanpy>=1.9.1
scvi-tools>=0.20.0
sphinx<6
statsmodels
protobuf~=3.19.0
# requirements-dev.txt
black==22.10.0
flake8
pre-commit
pydata-sphinx-theme
pytest
pytest-cov
pytest-xdist
pytype
